Extract conflict marker helpers partial
